Move `query` and `set` into `internal/`
