+++ Makefile
@@ -220,6 +220,7 @@ vet:
 
 lint: tools
 	golint -set_exit_status ./cmd/...
+	golint -set_exit_status ./internal/...
 	golint -set_exit_status ./pkg/...
 
 fmt:
