+++ pkg/cmd/cli/bios/set.go
@@ -25,9 +25,9 @@ OTHER DEALINGS IN THE SOFTWARE.
 package bios
 
 import (
+	"github.com/Cray-HPE/gru/internal/set"
 	"github.com/Cray-HPE/gru/pkg/cmd/cli"
 	"github.com/Cray-HPE/gru/pkg/cmd/cli/bios/collections"
-	"github.com/Cray-HPE/gru/pkg/set"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"github.com/stmcginnis/gofish/redfish"
