+++ pkg/cmd/cli/nics/show.go
@@ -27,9 +27,9 @@
 package nics
 
 import (
+	"github.com/Cray-HPE/gru/internal/query"
 	"github.com/Cray-HPE/gru/pkg/auth"
 	"github.com/Cray-HPE/gru/pkg/cmd/cli"
-	"github.com/Cray-HPE/gru/pkg/query"
 	"github.com/spf13/cobra"
 )
